label component removed from typography.tsx and moved to label.tsx
